Queue chunk update when chunk is busy
Issue #16 Bug where achunk is already in progress updating from a
previous change when a new change happens and the chunk is busy. The
first chunk update builds the mesh for the block with the missing face
before the change happens to remove it.
This change makes the second update wait for the first one to finish if
none are already waiting