Added texture index, change mesh block definitions
Added a texture index to replace the texture atlas and automatically
package all the textures in resources/textures into an atlas using
Kenney's code for this. The textures are added to an index based on the
file name (but anything after a "-" is ignored). Blocks definitions now
take the texture name, not the tile position and that finds the texture
collection for that texture. If you have multiple textures with the same
name before the dash (-) the blocks using that texture name will
randomly choose a texture of the ones available.  Textures can be
variable sizes so one can be 8x8 and another could be 128x128 for
example.

Mesh blocks are no longer defined in the block definition, instead
anything in resources/meshes is imported as a block using the mesh name
(Not always the same as the file name). These then use the texture by
the same name in the resources/textures folder. This is done at runtime.

I'm committing this to a branch because it seems to slow down game start
time and will break a lot of functionality for old textures and mesh
blocks.