Generate random seed based on world name
The question in Issue #28 (now closed) got me thinking that a random
seed shouldn't bee to hard to implement so I added this to generate a
seed for the noise class. This also makes Noise not static so things
need to maintain a reference to it. World sends the world name as the
array, the random textures just use the default because they don't
really need a unique layout and it costs a little time at startup to
generate a new permutation array.

This should really have been done in master but I did it here for some
reason. I'm going to need to merge this branch soon.