Add editor window and custom terrain
This will break old terrain generation BUT you can now define terrain
generation for the world in an editor window. Open it in Window ->
Voxelmetric, it also has a tab for defining blocks. Blocks can still be
defined the old way, this just puts a gui on top of it. Terrain layers
are also defined as components and their order is defined in the
TerrainGen.