HashMap for chunk data and flags
Added a hashmap to chunks so that you can store data on a per chunk
level and also moved all the boolean flags to there with a Flag enum as
the key. This can be extended to add almost anything to the flags with
SetFlag and GetFlag, GetFlag by default gets a boolean but you can
specify the type with GetFlag<Type>(key).