Don't autoload mesh blocks, use block definition
Mesh blocks are now added like other blocks and not auto loaded. They
also have several parameters, They can be solid on any side meaning
blocks on that side wont render the face facing it, there is an offset
that will be applied on import moving the mesh, texture name and mesh
name are for the texture and mesh files for this mesh. The mesh file is
now found by the file name NOT the submesh name and all the submeshes
are combined. The custom meshes now also generate colliders.