Change from local to global block positions
Wherever possible we now use global positions. Exceptions are indexes
for the blocks array and the render function for blocks. The render
function gets both local and global position. This is because the
vertices are placed according to the local position and because getting
a block from the chunk is slightly faster if you use a local position
instead of a global one so theres a LocalGetBlock function for just this
reason. This is not necessary for all GetBlock calls but in rendering so
many GetBlock calls are made that it slows down considerably when using
the global GetBlock function.