set server IP in inspector
Also looked into what it would take to reveal config structs like block
and texture configs in the inspector. It's feasible but needs a way to
recognize that the config already exists and not overwrite it by trying
to fetch the json version. Would probably also need to initialize things
like the textureIndex on init so they are accessible in the editor. Oh
well, not sure what the point would be anyway, I see no reason to edit
hundreds of blocks in a crappy editor window that works unreliably.