Use voxel based raycast function
Voxels no longer have col meshes so this voxel traversal algorithm is
needed to select a block with a raycast. I couldn't get A Fast Voxel
Traversal Algorithm for Ray Tracing by John Amanatides and Andrew Woo to
work with an origin not aligned to the centre of a voxel so I wrote my
own which is why it took longer than anticipated. If someone takes a
look at that thing to optimize / fix it I will be grateful.