Basic single threading support
Single threading is really slow now. It runs between 30 and 60 fps for
me but the chunk generation lags behind the camera position. I suspect
the problem was changing chunk generation to be done per chunk and not
per chunk column. It leads to a lot of extra time spent generating which
is covered up when it's in a thread but on the main thread it becomes
significant.